skip quasar warn log
